[ADD] speechRecognition util for Separation of concerns, redability
